bugfix
Note, this code is generated via an AWS JSON API file, so we can't just "merge this bugfix in". We need to fix the original string conversion library Rupert used. I'm just fixing this so I can get our project out the door and will come back to a longer term fix.